main: add DMA read ADC and FLASH r/w operation
The ADC data is read periodically by TIM4 which has an interval
of 500ms. When TIM4 trigger an interrupt (Counter Period reached)
we start DMA1 to go collect the ADC data. The Task1 has higher
interrupt priority than other Task, but it only process data when
DMA finished collecting the ADC data so most of the time it's asleep.

Task2 and Task3 do nothing for now, displaying LED2 and 3 to
aid in debugging.